refactor : 폴더 이동(mypage > profile)
