+++ src/pages/profile/index.tsx
@@ -3,8 +3,8 @@ import Layout from '@components/common/Layout'
 import Loader from '@components/common/Loader'
 import Navigate from '@components/common/Navigate'
 import Tab from '@components/common/Tab'
-import Activity from '@components/mypage/Activity'
-import SettingItem from '@components/mypage/SettingItem'
+import Activity from '@components/profile/Activity'
+import SettingItem from '@components/profile/SettingItem'
 import useGetProfile from '@hooks/queries/useGetProfile'
 import Image from 'next/image'
 import tw from 'tailwind-styled-components'
